metcher message payload and project dirs changes
