+++ src/common/Avatar.tsx
@@ -2,7 +2,7 @@ import { FC } from "react";
 import classnames from "classnames";
 
 import BotAvatar from "../assets/bot.svg";
-import { useMessageContext } from "../hooks";
+import { useMessageContext } from "src/messages/hooks";
 
 import classes from "./Avatar.module.css";
 interface AvatarProps {
