--- src/lib/css.ts
@@ -1,10 +1,0 @@
-export const getBackgroundImage = (url: string) => {
-	if (!url) return undefined;
-
-	const escapedUrl = url
-		.replace(/\n/g, "")
-		.replace(/\r/g, "")
-		.replace(/"\\/g, char => `\`${char}`);
-
-	return `url("${escapedUrl}")`;
-};
