+++ src/messages/List/List.tsx
@@ -1,6 +1,6 @@
 import { FC, Fragment, useEffect, useMemo } from "react";
 import ListItem from "./ListItem";
-import { useMessageContext } from "src/hooks";
+import { useMessageContext } from "src/messages/hooks";
 import mainclasses from "src/main.module.css";
 import classes from "./List.module.css";
 import classnames from "classnames";
