+++ src/messages/Text/Text.tsx
@@ -2,7 +2,7 @@ import { FC } from "react";
 import classes from "./Text.module.css";
 
 import ChatBubble from "../../common/ChatBubble";
-import { useMessageContext } from "../../hooks";
+import { useMessageContext } from "../hooks";
 
 interface TextProps {
 	content?: string;
