+++ src/messages/Video/Video.tsx
@@ -3,7 +3,7 @@ import ReactPlayer from "react-player";
 import classes from "./Video.module.css";
 import mainClasses from "src/main.module.css";
 import { VideoPlayIcon } from "src/assets/svg";
-import { useMessageContext } from "src/hooks";
+import { useMessageContext } from "src/messages/hooks";
 import { getChannelPayload } from "src/utils";
 
 const Video: FC = () => {
