+++ src/messages/index.ts
@@ -1,3 +1,8 @@
 import Text from "./Text";
+import Image from "./Image";
+import Video from "./Video";
+import Audio from "./Audio";
+import List from "./List";
+import TextWithButtons from "./TextWithButtons/TextWithButtons";
 
-export { Text };
+export { Text, Image, Video, Audio, List, TextWithButtons };
