+++ src/utils.ts
@@ -29,6 +29,17 @@ export const getWebchatButtonLabel: getWebchatButtonLabel = button => {
 	return title;
 };
 
+export const getBackgroundImage = (url: string) => {
+	if (!url) return undefined;
+
+	const escapedUrl = url
+		.replace(/\n/g, "")
+		.replace(/\r/g, "")
+		.replace(/"\\/g, char => `\`${char}`);
+
+	return `url("${escapedUrl}")`;
+};
+
 export const getRandomId = (prefix = "") => {
 	const id = window?.crypto?.randomUUID?.() || Date.now();
