+++ test/Audio.spec.tsx
@@ -1,6 +1,6 @@
 import { render, waitFor, screen } from "@testing-library/react";
 import { it, describe, expect } from "vitest";
-import Message from "src/Message";
+import Message from "src/messages/Message";
 import audio from "test/fixtures/audio.json";
 
 describe("Message Audio", () => {
