+++ test/Image.spec.tsx
@@ -1,6 +1,6 @@
 import { render, waitFor, screen } from "@testing-library/react";
 import { it, describe, expect } from "vitest";
-import Message from "src/Message";
+import Message from "src/messages/Message";
 import image from "test/fixtures/image.json";
 
 describe("Message Image", () => {
